+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/settings/TrackingProtectionFragmentTest.kt
@@ -52,6 +52,12 @@ class TrackingProtectionFragmentTest {
         val customRedirectTrackersCheckBox = settingsFragment.customRedirectTrackers.isChecked
         val customRedirectTrackersCheckBoxSettings = settings.blockRedirectTrackersInCustomTrackingProtection
 
+        val customSuspectedFingerprinters = settingsFragment.customSuspectedFingerprinters.isChecked
+        val customSuspectedFingerprintersSetting = settings.blockSuspectedFingerprintersInCustomTrackingProtection
+
+        val customSuspectedFingerprintersSelect = settingsFragment.customSuspectedFingerprintersSelect.value
+        val customSuspectedFingerprintersSelectSetting = settings.blockSuspectedFingerprintersSelectionInCustomTrackingProtection
+
         assertEquals(customCookiesCheckBoxSettings, customCookiesCheckBox)
         assertEquals(customCookiesSelectSettings, customCookiesSelect)
         assertEquals(customTrackingContentCheckBoxSettings, customTrackingContentCheckBox)
@@ -59,5 +65,7 @@ class TrackingProtectionFragmentTest {
         assertEquals(customCryptominersCheckBoxSettings, customCryptominersCheckBox)
         assertEquals(customFingerprintersCheckBoxSettings, customFingerprintersCheckBox)
         assertEquals(customRedirectTrackersCheckBoxSettings, customRedirectTrackersCheckBox)
+        assertEquals(customSuspectedFingerprinters, customSuspectedFingerprintersSetting)
+        assertEquals(customSuspectedFingerprintersSelect, customSuspectedFingerprintersSelectSetting)
     }
 }